+++ builtin/providers/aws/structure.go
@@ -1005,6 +1005,19 @@ func flattenAsgEnabledMetrics(list []*autoscaling.EnabledMetric) []string {
 	return strs
 }
 
+func flattenApiGatewayStageKeys(keys []*string) []map[string]interface{} {
+	stageKeys := make([]map[string]interface{}, 0, len(keys))
+	for _, o := range keys {
+		key := make(map[string]interface{})
+		parts := strings.Split(*o, "/")
+		key["stage_name"] = parts[1]
+		key["rest_api_id"] = parts[0]
+
+		stageKeys = append(stageKeys, key)
+	}
+	return stageKeys
+}
+
 func expandApiGatewayStageKeys(d *schema.ResourceData) []*apigateway.StageKey {
 	var stageKeys []*apigateway.StageKey