+++ builtin/provisioners/file/resource_provisioner_test.go
@@ -11,7 +11,7 @@ func TestResourceProvisioner_impl(t *testing.T) {
 	var _ terraform.ResourceProvisioner = new(ResourceProvisioner)
 }
 
-func TestResourceProvider_Validate_good(t *testing.T) {
+func TestResourceProvider_Validate_good_source(t *testing.T) {
 	c := testConfig(t, map[string]interface{}{
 		"source":      "/tmp/foo",
 		"destination": "/tmp/bar",
@@ -26,9 +26,40 @@ func TestResourceProvider_Validate_good(t *testing.T) {
 	}
 }
 
-func TestResourceProvider_Validate_bad(t *testing.T) {
+func TestResourceProvider_Validate_good_content(t *testing.T) {
+	c := testConfig(t, map[string]interface{}{
+		"content":     "value to copy",
+		"destination": "/tmp/bar",
+	})
+	p := new(ResourceProvisioner)
+	warn, errs := p.Validate(c)
+	if len(warn) > 0 {
+		t.Fatalf("Warnings: %v", warn)
+	}
+	if len(errs) > 0 {
+		t.Fatalf("Errors: %v", errs)
+	}
+}
+
+func TestResourceProvider_Validate_bad_not_destination(t *testing.T) {
 	c := testConfig(t, map[string]interface{}{
 		"source": "nope",
+	})
+	p := new(ResourceProvisioner)
+	warn, errs := p.Validate(c)
+	if len(warn) > 0 {
+		t.Fatalf("Warnings: %v", warn)
+	}
+	if len(errs) == 0 {
+		t.Fatalf("Should have errors")
+	}
+}
+
+func TestResourceProvider_Validate_bad_to_many_src(t *testing.T) {
+	c := testConfig(t, map[string]interface{}{
+		"source":      "nope",
+		"content":     "value to copy",
+		"destination": "/tmp/bar",
 	})
 	p := new(ResourceProvisioner)
 	warn, errs := p.Validate(c)