+++ terraform/graph_config_node_variable.go
@@ -179,6 +179,12 @@ func (n *GraphNodeConfigVariable) EvalTree() EvalNode {
 				VariableValues: variables,
 			},
 
+			&EvalCoerceMapVariable{
+				Variables:  variables,
+				ModulePath: n.ModulePath,
+				ModuleTree: n.ModuleTree,
+			},
+
 			&EvalTypeCheckVariable{
 				Variables:  variables,
 				ModulePath: n.ModulePath,