+++ terraform/test-fixtures/plan-module-map-literal/main.tf
@@ -1,0 +1,7 @@
+module "child" {
+  source = "./child"
+  amap {
+    foo = "bar"
+  }
+  othermap {}
+}